ISim style for active-hi signals in waveform
This means that instead of lo-high transition, whole range
of hi is filled with signal color. Spotted in Xilinx's ISim.
Without the fill, it is difficult to understand if signal is
1 or 0 when it is constant throughout the waveform window.

I also highlight selected signals by bold and selection can be
done easier - clicking any point in the tree row.