Word occurrence highlight for all editors. Merged their common parts.
Word search is purely text search, not even parser/ast is
used. It might be needed for Outline update, which currently
stays unconnected from editor selection also.

In fact, we could use AST search if doc was not edited
and relay on this seach when doc was edited. Outline
is built by AST anyway.